requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
discord==2.3.2
discord.py==2.4.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.0
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.135
MarkupSafe==3.0.2
marshmallow==3.22.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.1
numpy==1.26.4
ollama==0.3.3
orjson==3.10.7
packaging==24.1
pillow==10.4.0
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
